fix(tests): Move tests to correct package as source
All these tests were in com.netflix.spinnaker.echo.pipelinetriggers.
They are testing classes in
com.netflix.spinnaker.echo.pipelinetriggers.monitor.

This change moves the tests to
com.netflix.spinnaker.echo.pipelinetriggers.monitor as well to match
their corresponding source classes.
